Remove feature flags from CLP's
Feature flags require reliable information about current user and that
is not available in landing page ctrlr. Feature flags are used only for
google tag manager and are not relevant at this point to clps. If flags
are required to track e.g. AB testing in the future, this issue needs to
be solved properly.